refactor(hosted): satisfy robustness gates
