+++ lib/cluster-worker/engine-adapter.js
@@ -1,6 +1,7 @@
 'use strict';
 
 const { cloneJson } = require('./object-utils');
+const { inputFromRequest } = require('./engine-input');
 
 const TERMINAL_TOPICS = new Set(['CLUSTER_COMPLETE', 'CLUSTER_FAILED']);
 
@@ -30,17 +31,6 @@ function terminalEventFromMessage(message) {
     summary: data.reason || message.content?.text,
     code: data.code,
     reason: data.workerReason,
-  };
-}
-
-function inputFromRequest(request, artifactManifest) {
-  if (request.source === 'issue') return { issue: request.issue };
-  if (request.source === 'prompt') return { text: request.prompt };
-  return {
-    text:
-      'Execute the task described by this registry-staged, byte-free artifact manifest. ' +
-      'Do not request interactive input.\n' +
-      JSON.stringify(artifactManifest),
   };
 }
 
@@ -243,16 +233,16 @@ class CurrentEngineAdapter {
     this.stopPromise ||= this.stopResource();
     return this.stopPromise;
   }
+
   async waitForAllocatedCluster() {
     let cluster = this.orchestrator.getCluster(this.resource.clusterId);
     while (!cluster && !this.startSettled && !this.closed) {
-      await new Promise((resolve) => {
-        setTimeout(resolve, 10);
-      });
+      await new Promise((resolve) => setTimeout(resolve, 10));
       cluster = this.orchestrator.getCluster(this.resource.clusterId);
     }
     return cluster;
   }
+
   stopAllocatedCluster(cluster = this.orchestrator.getCluster(this.resource.clusterId)) {
     if (!cluster) return Promise.resolve(false);
     if (cluster.state === 'stopped' || cluster.state === 'killed') return Promise.resolve(true);
@@ -261,6 +251,7 @@ class CurrentEngineAdapter {
       .then(() => true);
     return this.allocatedStopPromise;
   }
+
   async stopResource() {
     this.cleanupPromise ||= this.waitForAllocatedCluster()
       .then((cluster) => this.stopAllocatedCluster(cluster))
